+++ app/controllers/ahoy_captain/exit_pages_controller.rb
@@ -14,7 +14,7 @@ module AhoyCaptain
                       .order(Arel.sql "count(#{AhoyCaptain.config.event[:url_column]}) desc")
                               .limit(limit)
       end
-      @pages = paginate(results).map { |page| ExitPageDecorator.new(page) }
+      @pages = paginate(results).map { |page| ExitPageDecorator.new(page, self) }
     end
   end
 end