+++ app/models/ahoy_captain/current.rb
@@ -1,9 +1,0 @@
-module AhoyCaptain
-  class Current < ActiveSupport::CurrentAttributes
-    attribute :request
-
-    after_reset do
-      self.request = nil
-    end
-  end
-end