+++ app/models/ahoy_captain/url_helpers.rb
@@ -1,6 +1,0 @@
-module AhoyCaptain
-  class UrlHelpers
-    include AhoyCaptain::Engine.routes.url_helpers
-    include Singleton
-  end
-end